Added limit on number of pre-fetched tokens
The token pre-fetch logic will now only fetch icons for the first few tokens. This is to prevent the possibility of memory issues if the user has hundreds of tokens with multi-MB images; only the first token images would be loaded, rather than all of them.

Currently, the number of fetched token icons is hard-coded to 12, which should be sufficient to load enough tokens to fill a tablet.