Add workaround for intermittent "Caller is not foreground"
Occasionally, when the unlock dialog should be shown, instead the user is directed to an empty token list page. None of the authentication callback methods are being called (either success or failure), and the ERROR-level message "Caller is not foreground" is logged by AuthService. Adding a brief (100 ms) wait before showing the biometric dialog resolves the issue, though this workaround is obviously terrible.